# workshop-4/python-temp/swagger.yml
openapi: 3.1.0
info:
  title: LBK Points - Transfer API
  version: 1.1.0
  description: |
    โอนแต้ม, ดูสถานะ, และค้นประวัติ พร้อม CRUD User management
servers:
  - url: http://localhost:3000
...
components:
  schemas:
    MembershipLevel:
      type: string
      enum: [BRONZE, SILVER, GOLD, PLATINUM]
...
    UserBase:
      type: object
      required:
        - first_name
        - last_name
        - email
      properties:
        first_name:
          type: string
          description: User's first name
          example: สมชาย
        last_name:
          type: string
          description: User's last name
          example: ใจดี
        phone:
          type: string
          nullable: true
          description: User's phone number
          example: "081-234-5678"
        email:
          type: string
          format: email
          description: User's email address (unique)
          example: somchai@example.com
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          required:
            - id
            - member_since
            - membership_level
            - points
            - created_at
            - updated_at
          properties:
            id:
              type: integer
              description: Unique identifier for the user
              example: 123
            member_since:
              type: string
              format: date-time
              description: Date when user became a member
            membership_level:
              $ref: '#/components/schemas/MembershipLevel'
            points:
              type: integer
              minimum: 0
              description: Current points balance
              example: 1500
            created_at:
              type: string
              format: date-time
              description: Record creation timestamp
            updated_at:
              type: string
              format: date-time
              description: Record last update timestamp
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            membership_level:
              $ref: '#/components/schemas/MembershipLevel'
            points:
              type: integer
              minimum: 0
              description: Current points balance
              example: 100
    UserUpdate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            membership_level:
              $ref: '#/components/schemas/MembershipLevel'
            points:
              type: integer
              minimum: 0
              description: Current points balance
            member_since:
              type: string
              format: date-time
              description: Date when user became a member
    ApiResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          description: Response status
          example: success
        message:
          type: string
          description: Response message
          example: Operation completed successfully
        data:
          description: Response data
    PointsOperation:
      type: object
      required:
        - points
      properties:
        points:
          type: integer
          description: Points amount
          example: 100
          minimum: 1
...
paths:
...
  /users:
    get:
      tags: [User Management]
      summary: Get all users
      description: Retrieve a list of all users from the SQLite database
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserResponse'
              example:
                status: success
                message: Users retrieved successfully
                data:
                  - id: 1
                    first_name: สมชาย
                    last_name: ใจดี
                    phone: 081-234-5678
                    email: somchai@example.com
                    member_since: "2024-01-15T10:30:00"
                    membership_level: GOLD
                    points: 1500
                    created_at: "2024-01-15T10:30:00"
                    updated_at: "2024-01-15T10:30:00"
    
    post:
      tags: [User Management]
      summary: Create a new user
      description: Create a new user in the SQLite database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              complete:
                summary: Example with all user fields
                value:
                  first_name: สมชาย
                  last_name: ใจดี
                  phone: 081-234-5678
                  email: somchai@example.com
                  membership_level: SILVER
                  points: 100
              minimal:
                summary: Example with required fields only
                value:
                  first_name: สมศรี
                  last_name: สุขใจ
                  email: somsri@example.com
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data or email already exists
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: null
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: User's unique identifier
        schema:
          type: integer
          format: int64
          example: 1
    
    get:
      tags: [User Management]
      summary: Get user by ID
      description: Retrieve a specific user by their unique identifier from SQLite database
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    put:
      tags: [User Management]
      summary: Update an existing user
      description: Update user information for an existing user in SQLite database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            example:
              first_name: สมชาย
              last_name: ใจดี (แก้ไข)
              phone: 081-234-5678
              email: somchai.updated@example.com
              membership_level: GOLD
              points: 2000
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      tags: [User Management]
      summary: Delete a user
      description: Delete a user from the SQLite database by their ID
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{id}/points/add:
    parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          example: 1
    
    post:
      tags: [User Management]
      summary: Add points to user
      description: Add points to a user's account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointsOperation'
      responses:
        '200':
          description: Points added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{id}/points/deduct:
    parameters:
      - name: id
        in: path
        required: true
        description: User ID
        schema:
          type: integer
          example: 1
    
    post:
      tags: [User Management]
      summary: Deduct points from user
      description: Deduct points from a user's account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointsOperation'
      responses:
        '200':
          description: Points deducted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '400':
          description: Insufficient points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...